tests/wpt/web-platform-tests/html/semantics/grouping-content/the-dd-element/grouping-dd.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>the dd element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="https://html.spec.whatwg.org/multipage/#the-dd-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        "use strict";

        // check that prototype matches spec's DOM interface
        test(function () {
            var testElement = document.createElement("dd");
            assert_equals(Object.getPrototypeOf(testElement), HTMLElement.prototype, "HTMLElement.prototype should be used for dd");
        }, "The prototype for dd is HTMLElement.prototype");

    </script>
</head>
<body>
    <h1>Description</h1>
    <p>This test validates the dd element.</p>

    <div id="log"></div>
</body>
</html>
